# poc2-tuya-dotnet-sdd/specs/001-create-query-orders/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Orders API
  version: 1.0.0
  description: Create valid orders and retrieve them later by exact identifier.
  license:
    name: Internal proof of concept
    identifier: LicenseRef-Internal-PoC
servers:
  - url: /
security: []
paths:
  /orders:
    post:
      operationId: createOrder
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          headers:
            Location:
              description: Relative URL of the created order.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid order data, including a duplicate product.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ValidationProblem'
        '500':
          description: Unexpected server error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /orders/{orderId}:
    get:
      operationId: getOrder
      summary: Retrieve an order by exact identifier
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid order identifier.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ValidationProblem'
        '404':
          description: Order not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  schemas:
    CreateOrderRequest:
      type: object
      additionalProperties: false
      required:
        - customerId
        - items
      properties:
        customerId:
          type: string
          description: Opaque identifier containing at least one non-whitespace character.
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateOrderItem'
    CreateOrderItem:
      type: object
      additionalProperties: false
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
          description: Opaque identifier containing at least one non-whitespace character.
        quantity:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
    Order:
      type: object
      additionalProperties: false
      required:
        - id
        - customerId
        - createdAt
        - items
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderItem:
      type: object
      additionalProperties: false
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
        quantity:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
    Problem:
      type: object
      additionalProperties: true
      required:
        - title
        - status
        - traceId
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        traceId:
          type: string
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          required:
            - errors
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string